Updated pronunciation in UX element

+++ DFC.App.JobGroups.Data/Models/JobGroupModels/JobGrowthPredictionModel.cs
@@ -9,11 +9,11 @@ namespace DFC.App.JobGroups.Data.Models.JobGroupModels
 
         public int EndYearRange { get; set; }
 
-        public decimal JobsCreated { get; set; }
+        public int JobsCreated { get; set; }
 
         public decimal PercentageGrowth { get; set; }
 
-        public decimal? Retirements { get; set; }
+        public int? Retirements { get; set; }
 
         public decimal? PercentageRetirements { get; set; }
     }
